Add AWS permissions for Kops 1.21
If this commit applied nodes will have the permissions that new version of Kops 1.21 needs to create a new cluster.
We hit the limit of policies that can be attached so also repurposed mattermost-provisioner-ec2-policy to include any new categories needed that we don't have a specific policy for them already attached.
Issue: MM-37151