Add a new lambda functions called EBS janitor
The EBS janitor is responsible to delete volumes which are not in use and they have been
created very long time ago (specifically default configuration is 90 days). If a volume
is available and created more than 90days (~3 months) we delete the volume for cleaning up
and keep the cost down.

Ticket: https://mattermost.atlassian.net/browse/MM-37523