[vpc-setup] Add an extra policy for velero permissions (#413)
This is a temporary solution we really need to invest time to have one single point of
infrastructure as a code for cluster creation. Right now is not clear where is this going
to be attached but in practice is attached in Provisioner codebase. Check this file and
specifically this line [here](https://github.com/mattermost/mattermost-cloud/blob/master/internal/provisioner/kops_provisioner_cluster.go#L208).
There are extra lines like this.

Apart from this we need to invest time to move away from node policies and move all of
our services to kube2iam.

Ticket: https://mattermost.atlassian.net/browse/CLD-998